[AIRFLOW-5786] Move into updated destination ref AIP-21 2019-11-05
