+++ airflow/contrib/hooks/aws_sns_hook.py
@@ -17,15 +17,15 @@
 # specific language governing permissions and limitations
 # under the License.
 
-"""This module is deprecated. Please use `airflow.providers.aws.hooks.sns`."""
+"""This module is deprecated. Please use `airflow.providers.amazon.aws.hooks.sns`."""
 
 import warnings
 
 # pylint: disable=unused-import
-from airflow.providers.aws.hooks.sns import AwsSnsHook  # noqa
+from airflow.providers.amazon.aws.hooks.sns import AwsSnsHook  # noqa
 
 warnings.warn(
-    "This module is deprecated. Please use `airflow.providers.aws.hooks.sns`.",
+    "This module is deprecated. Please use `airflow.providers.amazon.aws.hooks.sns`.",
     DeprecationWarning,
     stacklevel=2,
 )
