+++ airflow/providers/amazon/aws/operators/sns.py
@@ -20,7 +20,7 @@
 """Publish message to SNS queue"""
 
 from airflow.models import BaseOperator
-from airflow.providers.aws.hooks.sns import AwsSnsHook
+from airflow.providers.amazon.aws.hooks.sns import AwsSnsHook
 from airflow.utils.decorators import apply_defaults
